+++ src/Koffee/MainController.fs
@@ -160,7 +160,7 @@ type Controller(
             keyHandler.Handle()
             let handler =
                 match confirmType with
-                | Overwrite (putType, srcExistingPairs) -> itemActionHandler.ConfirmOverwrite putType srcExistingPairs
+                | Overwrite _ -> itemActionHandler.ConfirmOverwrite
                 | Delete items -> itemActionHandler.ConfirmDelete items
                 | OverwriteBookmark (char, _) -> navigationHandler.ConfirmOverwriteBookmark char
                 | OverwriteSavedSearch (char, _) -> navigationHandler.ConfirmOverwriteSavedSearch char